Ändern Funktion halbwegs eingebunden
Flo

+++ src/de/dpma/view/MainPageController.java
@@ -1,6 +1,5 @@
 package de.dpma.view;
 
-import java.io.File;
 import java.sql.SQLException;
 
 import de.dpma.FXML_GUI;
@@ -12,109 +11,122 @@ import de.dpma.model.Dozent;
 import de.dpma.model.Event;
 import de.dpma.model.Stundenlohn;
 import de.dpma.util.AlertUtil;
-import de.dpma.util.WriteDocxTEST;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
 import javafx.scene.control.ListView;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
-import javafx.stage.FileChooser;
 
 public class MainPageController {
-	
+
 	@FXML
 	ListView<String> navigationListe = new ListView<String>();
-	
+
 	@FXML
 	TableView tabellenTableView;
-	
+
 	AlertUtil alert;
-	
+
 	RootLayoutController root = new RootLayoutController();
-	
+
 	String fokus = "Veranstaltungen";
-	
+
 	static DozentDAO dozentDAO = new DozentDAO(MainApp.dbcon.getConnection());
-	
+
 	static EventDAO eventDAO = new EventDAO(MainApp.dbcon.getConnection());
-	
+
 	static StundenlohnDAO stundenlohnDAO = new StundenlohnDAO(MainApp.dbcon.getConnection());
-	
+
 	Event event;
-	
+
 	Dozent dozent;
-	
+
 	Stundenlohn stundenlohn;
-	
+
 	@FXML
 	public void initialize() {
-		
+
 		FXML_GUI.primaryStage.setTitle(fokus);
 		ObservableList<String> inhalte = FXCollections.observableArrayList("Veranstaltungen", "Dozenten",
 				"Lehrvergütungssätze");
 		navigationListe.setItems(inhalte);
-		
+
 		handleVeranstaltungen();
-		
+
 	}
-	
+
 	@FXML
 	public void handleSelect() {
-		
+
 		fokus = navigationListe.getFocusModel().getFocusedItem();
-		
+
 		FXML_GUI.primaryStage.setTitle(fokus);
-		
+
 		if (fokus.equals("Veranstaltungen")) {
 			handleVeranstaltungen();
-		}
-		else if (fokus.equals("Dozenten")) {
+		} else if (fokus.equals("Dozenten")) {
 			handleDozenten();
-		}
-		else if (fokus.equals("Lehrvergütungssätze")) {
+		} else if (fokus.equals("Lehrvergütungssätze")) {
 			handleLehrverguetung();
-		}
-		else {
+		} else {
 			alert = new AlertUtil("Der Ausgewählte Punkt ist ungültig", "Bitte kontaktieren Sie den Administrator",
 					"WARNING");
 		}
 	}
-	
+
 	@FXML
 	public void handleNew() {
-		
+
 		root.handleGUI(fokus, null);
 	}
-	
+
 	@FXML
 	public void handleDelete() {
-		
+
 		int selectedIndex = tabellenTableView.getSelectionModel().getSelectedIndex();
-		if (fokus.equals("Veranstaltungen")) {
-			event = (Event) tabellenTableView.getSelectionModel().getSelectedItem();
-			tabellenTableView.getItems().remove(selectedIndex);
-			try {
-				eventDAO.deleteEvent(event.getId());
+		if (selectedIndex >= 0) {
+			if (fokus.equals("Veranstaltungen")) {
+				event = (Event) tabellenTableView.getSelectionModel().getSelectedItem();
+				tabellenTableView.getItems().remove(selectedIndex);
+
+				try {
+					eventDAO.deleteEvent(event.getId());
+				} catch (SQLException e) {
+					e.printStackTrace();
+				}
+			} else if (fokus.equals("Dozenten")) {
+				dozent = (Dozent) tabellenTableView.getSelectionModel().getSelectedItem();
+				tabellenTableView.getItems().remove(selectedIndex);
+
+				try {
+					dozentDAO.deleteDozent(dozent.getId());
+				} catch (SQLException e) {
+					e.printStackTrace();
+				}
+			} else if (fokus.equals("Lehrvergütungssätze")) {
+				stundenlohn = (Stundenlohn) tabellenTableView.getSelectionModel().getSelectedItem();
+				tabellenTableView.getItems().remove(selectedIndex);
+
+				try {
+					stundenlohnDAO.deleteStundenlohn(stundenlohn.getId());
+				} catch (SQLException e) {
+					e.printStackTrace();
+				}
 			}
-			catch (SQLException e) {
-				e.printStackTrace();
-			}
+		} else {
+			alert = new AlertUtil("Keine Auswahl",
+					"Sie haben kein zu löschendes Element ausgewählt. Bitte wählen Sie ein Element aus und versuchen Sie es erneut.",
+					"INFO");
 		}
-		else if (fokus.equals("Dozenten")) {
-			dozent = (Dozent) tabellenTableView.getSelectionModel().getSelectedItem();
-		}
-		else if (fokus.equals("Lehrvergütungssätze")) {
-			stundenlohn = (Stundenlohn) tabellenTableView.getSelectionModel().getSelectedItem();
-		}
-		
+
 		// aufrufen der SQL Befehles um den Eintrag zu droppen, dann kann die
 		// Tabelle nochmal aktuallisiert werden
 	}
-	
+
 	@FXML
-	public void handleEdit() {
-		
+	public void handleEdit() throws SQLException {
+
 		int selectedIndex = tabellenTableView.getSelectionModel().getSelectedIndex();
 		if (selectedIndex >= 0) {
 			switch (fokus) {
@@ -122,6 +134,7 @@ public class MainPageController {
 				event = (Event) tabellenTableView.getSelectionModel().getSelectedItem();
 				VeranstaltungController vc = new VeranstaltungController();
 				vc.handleNew(event);
+				root.handleGUI(fokus, event);
 				break;
 			case "Dozenten":
 				dozent = (Dozent) tabellenTableView.getSelectionModel().getSelectedItem();
@@ -133,17 +146,21 @@ public class MainPageController {
 				LehrverguetungssaetzeController lc = new LehrverguetungssaetzeController();
 				lc.handleNew(stundenlohn);
 				break;
-			
+
 			default:
 				break;
 			}
+		} else {
+			alert = new AlertUtil("Keine Auswahl",
+					"Sie haben kein zu bearbeitendes Element ausgewählt. Bitte wählen Sie ein Element aus und versuchen Sie es erneut.",
+					"INFO");
 		}
-		
-		root.handleGUI(fokus, null);
+
+		// root.handleGUI(fokus, null);
 	}
-	
+
 	private void handleVeranstaltungen() {
-		
+
 		TableColumn<Event, String> nameTableColumn = new TableColumn("Name");
 		TableColumn<Event, String> aktenzeichenTableColumn = new TableColumn("Aktenzeichen");
 		TableColumn<Event, String> schulArt = new TableColumn("SchulArt");
@@ -154,17 +171,17 @@ public class MainPageController {
 		TableColumn<Event, String> stdZahlTableColumn = new TableColumn("StdZahl");
 		TableColumn<Event, String> betragTableColumn = new TableColumn("Betrag");
 		TableColumn<Event, String> betrag_ABCTableColumn = new TableColumn("Betrag_ABC");
-		
+
 		tabellenTableView.getColumns().setAll(nameTableColumn, aktenzeichenTableColumn, schulArt, vfgTableColumn,
 				vortragTableColumn, datumTableColumn, euro_StdTableColumn, stdZahlTableColumn, betragTableColumn,
 				betrag_ABCTableColumn);
-		
+
 		try {
 			ObservableList<Event> eventData = FXCollections.observableArrayList();
 			eventData = FXCollections.observableArrayList((eventDAO.selectAllEvents()));
-			
+
 			tabellenTableView.setItems(eventData);
-			
+
 			nameTableColumn.setCellValueFactory(cellData -> cellData.getValue().DozentProperty());
 			aktenzeichenTableColumn.setCellValueFactory(cellData -> cellData.getValue().AktenzProperty());
 			schulArt.setCellValueFactory(cellData -> cellData.getValue().SchulartProperty());
@@ -175,18 +192,17 @@ public class MainPageController {
 			stdZahlTableColumn.setCellValueFactory(cellData -> cellData.getValue().getStundenZahlString());
 			betragTableColumn.setCellValueFactory(cellData -> cellData.getValue().BetragProperty());
 			betrag_ABCTableColumn.setCellValueFactory(cellData -> cellData.getValue().Betrag_ABCProperty());
-		}
-		catch (SQLException e) {
+		} catch (SQLException e) {
 			alert = new AlertUtil("Datenbankfehler",
-					"Es ist ein Fehler bei der Datenbankabfrage enstanden, Bitte kontaktieren Sie den Administrator",
+					"Es ist ein Fehler bei der Datenbankabfrage enstanden. Bitte kontaktieren Sie den Administrator",
 					"WARNING");
 			e.printStackTrace();
 		}
-		
+
 	}
-	
+
 	private void handleDozenten() {
-		
+
 		TableColumn<Dozent, String> anredeTableColumn = new TableColumn("Anrede");
 		TableColumn<Dozent, String> titelTableColumn = new TableColumn("Titel");
 		TableColumn<Dozent, String> vornameTableColumn = new TableColumn("Vorname");
@@ -200,14 +216,14 @@ public class MainPageController {
 		tabellenTableView.getColumns().setAll(anredeTableColumn, titelTableColumn, vornameTableColumn, nameTableColumn,
 				straßeTableColumn, pLZTableColumn, ortTableColumn, kontonummerTableColumn, bankTableColumn,
 				bLZTableColumn);
-		
+
 		try {
 			ObservableList<Dozent> dozentData = FXCollections.observableArrayList();
-			
+
 			dozentData = FXCollections.observableArrayList(dozentDAO.selectAllDozenten());
-			
+
 			tabellenTableView.setItems(dozentData);
-			
+
 			anredeTableColumn.setCellValueFactory(cellData -> cellData.getValue().AnredeProperty());
 			titelTableColumn.setCellValueFactory(cellData -> cellData.getValue().TitelProperty());
 			vornameTableColumn.setCellValueFactory(cellData -> cellData.getValue().VornameProperty());
@@ -218,79 +234,51 @@ public class MainPageController {
 			kontonummerTableColumn.setCellValueFactory(cellData -> cellData.getValue().IBANProperty());
 			bankTableColumn.setCellValueFactory(cellData -> cellData.getValue().BankProperty());
 			bLZTableColumn.setCellValueFactory(cellData -> cellData.getValue().BLZProperty());
-			
-		}
-		catch (SQLException e) {
+
+		} catch (SQLException e) {
 			e.printStackTrace();
 		}
-		
+
 	}
-	
+
 	private void handleLehrverguetung() {
-		
+
 		TableColumn<Stundenlohn, String> vergütungTableColumn = new TableColumn("Vergütung");
 		tabellenTableView.getColumns().setAll(vergütungTableColumn);
-		
+
 		try {
 			ObservableList<Stundenlohn> stundenlohnData = FXCollections.observableArrayList();
-			
+
 			stundenlohnData = FXCollections.observableArrayList(stundenlohnDAO.selectAllStundenloehne());
-			
+
 			tabellenTableView.setItems(stundenlohnData);
-			
+
 			vergütungTableColumn.setCellValueFactory(cellData -> cellData.getValue().LohnProperty());
-			
-		}
-		catch (SQLException e) {
+
+		} catch (SQLException e) {
 			e.printStackTrace();
 		}
-		
+
 	}
-	
+
 	public void handleCreateDocRechnung() {
-		
+
+		root.handleGUI("createDoc", "");
+
 		int selectedIndex = tabellenTableView.getSelectionModel().getSelectedIndex();
 		if (selectedIndex >= 0) {
-			switch (fokus) {
-			case "Veranstaltungen":
-				event = (Event) tabellenTableView.getSelectionModel().getSelectedItem();
-				break;
-			case "Dozenten":
-				
-				break;
-			case "Lehrvergütungssätze":
-				
-				break;
-			
-			default:
-				break;
-			}
+			event = (Event) tabellenTableView.getSelectionModel().getSelectedItem();
 		}
-		
-		try {
-			
-			// TODO Word export
-			FileChooser chooser = new FileChooser();
-			chooser.setTitle("Speichern");
-			FileChooser.ExtensionFilter extFilter = new FileChooser.ExtensionFilter("Word Datei (*.docx)", "*.docx");
-			chooser.getExtensionFilters().add(extFilter);
-			
-			String directory = chooser.getExtensionFilters().toString();
-			
-			File file = chooser.showSaveDialog(FXML_GUI.primaryStage.getScene().getWindow());
-			WriteDocxTEST wdoc = new WriteDocxTEST(file);
-		}
-		catch (Exception e) {
-		}
+
 	}
-	
+
 	public void handleCreateDocAuszahlung() {
-		
+
 		// TODO Word export
 	}
-	
+
 	public void handleSearch() {
-		
+
 		// TODO Suchenfunktion
 	}
 }
